added error checking and improvements to delete user form; corrected add user

+++ XPS/XPS/Forms/DeleteUserForm.cs
@@ -21,7 +21,7 @@ namespace XPS.Forms
         public DeleteUserForm(User user)
         {
             adminUser = user;
-            //DeleteUserForm.Text = user.FirstName + " " + user.LastName;
+            
             InitializeComponent();
         }
 
@@ -39,16 +39,84 @@ namespace XPS.Forms
 
         private void dGetUserButton_Click(object sender, EventArgs e)
         {
-            thisOne = new DatabaseManager();
-            deleteMe = thisOne.GetUser(Int32.Parse(dMustangBox.Text));
-            dFirstnameBox.Text = deleteMe.FirstName;
-            dLastnameBox.Text = deleteMe.LastName;
+
+            try
+            {
+                thisOne = new DatabaseManager();
+                deleteMe = thisOne.GetUser(Int32.Parse(dMustangBox.Text));
+                dFirstnameBox.Text = deleteMe.FirstName;
+                dLastnameBox.Text = deleteMe.LastName;
+                dInstructionBox.Hide();
+                dVerifyBox.Show();
+                dDeleteUserButton.Show();
+                dFirstnameLabel.Show();
+                dFirstnameBox.Show();
+                dLastnameLabel.Show();
+                dLastnameBox.Show();
+                dMustangBox.Text = "";
+               
+            }
+            catch(Exception ex)
+            {
+                MessageBox.Show("No record found for that Mustang ID.");
+                dMustangBox.Text = "";
+            }
+
+            
+            
         }
 
         private void dDeleteUserButton_Click(object sender, EventArgs e)
         {
-            MessageBox.Show("You are about to delete " + deleteMe.FirstName + " " + deleteMe.LastName);
-            thisOne.DeleteUser(deleteMe.UserID);
+           
+            DialogResult result1 = MessageBox.Show("Do you want to delete " +
+                deleteMe.FirstName + " " + deleteMe.LastName + "?",
+                "About to delete", MessageBoxButtons.YesNo);
+            if (result1 == DialogResult.Yes)
+            {
+                try
+                {
+                    bool deleteSuccessful = thisOne.DeleteUser(deleteMe.UserID);
+                    if (deleteSuccessful)
+                    {
+                        MessageBox.Show(deleteMe.FirstName +
+                       " " + deleteMe.LastName + " deleted from users table.");
+
+                        dMustangBox.Text = "";
+                        dFirstnameBox.Text = "";
+                        dLastnameBox.Text = "";
+                        dInstructionBox.Show();
+                        dVerifyBox.Hide();
+                        dDeleteUserButton.Hide();
+                        dFirstnameLabel.Hide();
+                        dFirstnameBox.Hide();
+                        dLastnameLabel.Hide();
+                        dLastnameBox.Hide();
+                    }
+                }
+                catch (Exception ex) { }
+            }
+            else
+            {
+                dMustangBox.Text = "";
+                dFirstnameBox.Text = "";
+                dLastnameBox.Text = "";
+            }
+
+            //MessageBox.Show("You are about to delete " + deleteMe.FirstName + " " + deleteMe.LastName);
+
+            //thisOne.DeleteUser(deleteMe.UserID);
+            /*
+            dFirstnameBox.Text = "";
+            dLastnameBox.Text = "";
+            dInstructionBox.Show();
+            dVerifyBox.Hide();
+            dDeleteUserButton.Hide();
+            dFirstnameLabel.Hide();
+            dFirstnameBox.Hide();
+            dLastnameLabel.Hide();
+            dLastnameBox.Hide();
+            */
         }
 
         private void dCancelButton_Click(object sender, EventArgs e)
